refactor: 당첨확인 로직 Lottos로 이동
