+++ src/main/java/lotto/domain/Lottos.java
@@ -1,10 +1,11 @@
 package lotto.domain;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 public class Lottos {
     private List<Lotto> lottos = new ArrayList<>();
+    private final Map<WinType, Integer> winResults = new EnumMap<>(WinType.class);
+    private static final int ZERO = 0;
 
     public Lottos() {
     }
@@ -25,16 +26,44 @@ public class Lottos {
         return new Lottos(manualLottos);
     }
 
-    public int size() {
-        return lottos.size();
+    public Map<WinType, Integer> matchWinning(WinningLotto winningLotto) {
+
+        initializeWinResults();
+
+        for (Lotto lotto : lottos) {
+            int matchCount = winningLotto.matchNumberCnt(lotto);
+            boolean matchBonus = winningLotto.matchBonus(lotto);
+            int existsWinCount = getWinCount(matchCount, matchBonus);
+            winResults.put(WinType.valueOf(matchCount, matchBonus), existsWinCount + 1);
+        }
+
+        return winResults;
     }
 
-    public List<Lotto> getLottos() {
-        return lottos;
+    public int size() {
+        return lottos.size();
     }
 
     public Lottos conbine(Lottos lottos) {
         this.lottos.addAll(lottos.getLottos());
         return new Lottos(this.lottos);
+    }
+
+    private int getWinCount(int matchCnt, boolean matchBonus) {
+        Integer winCount = winResults.get(WinType.valueOf(matchCnt, matchBonus));
+
+        if (winCount == null) {
+            return ZERO;
+        }
+        return winCount;
+    }
+
+    private void initializeWinResults() {
+        Arrays.stream(WinType.values())
+                .forEach(it -> winResults.put(it, ZERO));
+    }
+
+    public List<Lotto> getLottos() {
+        return lottos;
     }
 }
